change constant name to FLOAT

+++ parsons/databases/redshift/constants.py
@@ -1,6 +1,6 @@
 COL_NAME_MAX_LEN = 120
 
-DECIMAL = 'float'
+FLOAT = 'float'
 
 REPLACE_CHARS = {" ": ""}
